[win-bootstrap] Only check if git dirs are _within_depot_tools
Currently all dirs in PATH are checked for whether they are
within depot_tools, up to and including the git install dir.
It is expected that there will only be one depot_tools entry in PATH,
and so this would skip at most 2 os.path.isfile calls.

It makes more sense to limit the _within_depot_tools call to
directories that have a git command only, as this is expected to
be very few of the dirs in PATH.

Bug: b/360206460
Change-Id: I516f770aa2b59318231afd893f2cb17d5d9bef57