Add a --no-squash argument to `git rebase-update`
In common workflows, this step only succeeds ~5% of the time, and
in the other 95% of cases, simply adds ~30 seconds of wasted time
to the execution time of `rebase-update`. This CL doesn't change
the default behavior, but adds a way for users to opt out of this
wasted time.

Note that a previous attempt to land this [1] was reverted due
to the need to maintain a new flag. However, a subsequent CL
landed most of the needed functionality, which avoids squashing
when there is only a single commit in the branch. I believe the
maintenance burden added by this CL should be minimal, and the
time savings is very real. I've mostly written this CL during the
time I've been waiting for a rebase-update to complete.

[1] https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/2904096

Fixed: 40264739
Change-Id: Ib2a3ffe4b0e5d0b74323a2a0d34ab0e1b7fafb08