add yapf_format.py
YAPF requires .yapfignore to be present in the same folder where
yapf is executed. However, in chromium, the .yapfignore is located
in the top-level folder only, and _GetYapfIgnorePatterns() is made
to apply the top-level .yapfignore, whereever the binary is executed.
: https://source.chromium.org/chromium/chromium/tools/depot_tools/+/main:git_cl.py;l=738-765;drc=88bc812650299ff2ae07af2ff8d8731be8a448ba

This CL is intended to remove the function in git_cl.py.
Find the child CL for the change.

Bug: 356504967
Change-Id: I350aec31b27e9fede28b73022fe9b902b7eb7a16