Add `allow_reciprocal_licenses`
Add `allow_reciprocal_licenses` parameter to the metadata library and associated functions to enable open source license checks.

This parameter will allow for the use of reciprocal licenses for open source projects. By default, reciprocal licenses are not allowed.

This CL is a part of an effort to add support for license validation for Chromium's dependency metadata.


Bug: 385020146
Change-Id: I0374658207bc87ffd74e033762ee4973c6e83b3b