[git] Remove usage of recipe_util.format_ex
This function was a shim to ensure consistency between Python 2 and 3.
It's no longer necessary now that recipes don't support Python 2
anymore, and is being deleted in crrev.com/c/6133284.

Also update two tests to be less sensitive to the exact format of the
summary markdown produced when an exception occurs, since some single
quotes are now changed to double quotes.

Change-Id: Ia8a7bd9be5f03302049f6798c51b8e4e8bdc9784