Fix caching Gerrit account presence outside of repos
This fixes the case where a user tries to fetch new repos but isn't in
a Git directory (probably true for most people if they don't have,
e.g., their home directory as a Git repo).

There are a couple of ways to bikeshed a solution for this, but the
most straightforward/easiest is to skip since we're just caching the
account presence check.

Bug: b/328682976
Change-Id: I87a0b210451e6952839eef73616d012c41ad300f