split_cl: Make description file optional during dry runs
I frequently find myself running `git cl split` just to see how the files
are going to be split up, without actually intending to submit the resulting
CLs immediately. Doing so requires creating a file, putting a dummy description
in it, and passing that to the command line, which is unecessary work. This CL
simply allows dry runs to use a dummy description if none is provided, to make
checking the tool's output easier.

Bug: None
Change-Id: I47e06c6e6da26701e07dcae81ab605edac2e2ca6