Fix computing diffs in git_cl.py
The modified lines are corresponding to the following codes in
clang-format-diff.py
: https://source.chromium.org/chromium/chromium/src/+/main:third_party/clang-format/script/clang-format-diff.py;l=124-135

This CL fixes the following two issues.

1) if lines are simply removed w/o replacements in the original text,
_ComputeFormatDiffLineRanges() returns values that results in
--line $start:$start, which formats the start line number of the secion
removed and results in the line to be formatted. If lines are simply
removed, no formatting is necessary in the new text.

2) incorrect line counts in the next text.
The line count includes the starting line. Therefore, if the hunk has
12:1, it corresponds to the single, 12th line. However, the existing logic
translates it as 12th and 13th lines.

In most cases, the above issues won't make visible changes. Even if
they do, engineers wouldn't notice, as it would just format one additional
line below the changes that they made.

However, this CL is to match the same behaviour implmented in
clang-format-diff.py


Change-Id: Ifd2a19c8801c836f8e0799410b372a0f258fd476
Bug: 379902295