Presubmit error for bad delimiters
This change introduces a new error for license fields that use any
of the following `["/", ";", " and ", " or "]`.

I chose to include the offending character/s in the error message because
I find it easier to parse error messages that tell me exactly which 
character is the bad one. Similarly I've included conditions in the 
reason to handle the plural cases correctly, generating either:

`License contains a bad delimiter character {util.quoted(bad_values_in_license)}.`, or
`License contains bad delimiter characters {util.quoted(bad_values_in_license)}.`

I realise this means that any downstream rules looking to detect this error will need to check for a common subset, e.g 'bad delimiter
character', however I think it's worth it for the improved user
experience of receiving the error.

I've also anticipated that most of these errors will be due to situations
where multiple licenses are offered, and included additional text
explaining that only the most permissive of the choices should
be included.

Change-Id: I6eb53a8a3bd541a1801dff133884b719dcdfe04d