roll_dep: Update README.chromium when rolling submodules
Attempt to update the README.chromium file with the new submodule revision after rolling it in a DEPS update. This update is flag-guarded and will only run when explicitly specified.

Currently, only README.chromium files with a single "Revision:" line are supported. Multiple lines and files with the divider are not handled. These are left as future TODOs.

Change-Id: Ib776564ae94360cc72dd633fc7ed7b3f84b5b9d2