[dependency_metadata] Allow descriptions for CVEs
This adds a new way to report CVEs that includes an accompanying
description. It also adds a new validation check that ensures that the
CVE description is present for every entry listed in the 'Mitigated:'
field.

The 'mitigated' custom field library now exports the vulnerability regex
in two flavours: with anchors, and without. This is so that the parser
can extract them into self._entries, and then later in
dependency_metadata, these values can be extracted for aggregation and
validation.

## New Warnings added Warnings will now be created if cves (or other
vuln ids) are not uniformly provided in both the "Mitigated" list, as
well as in their own entries.

Bug: b/392026683
Change-Id: Ie55595970b49d705ac532f1f8c41ff47d959f56c