`git cl split`: Include dirname as part of branch names
This CL changes the way `git cl split` generates branch names during upload.
Specifically, it attempts to derive a common directory for all the files in
each CL, and includes that dirname as part of the branch name, in addition to
a hash of the files in the CL. This makes it a bit easier to guess what's in
a branch, given its name.

Since the hash is still included, branch names should remain unique. Finding a
common directory should be deterministic, so we can still rely on branches
having the same name if the script is re-run with the same splitting.

Bug: 389069356
Change-Id: I70490258755f13962f3db5c835619caa24176af9