`git cl split`: Save and load splittings from files
This CL implements the proposal (from the design doc in the cited bug) to
allow `git cl split` to save the generated splitting to a human-readable
and editable text file. This allows users to save the splitting in case
something goes wrong and they need to retry, and also lets them tweak
a generated splitting if they get something mostly, but not quite, usable.

Note that not all the suggestions from the doc have been incorporated yet,
but the bulk of the code is suitable for feedback.

Commit: False
Bug: 389069356
Change-Id: I8ae21f2eb7c022ba181ae5af273ce09605b5acec