git_cl: sets opts.full if format-full-by-default is set
settings.GetFormatFullByDefault() looks up a git config to find if
format-full-by-default is set.

RunClangFormatDiff() and RunGoogleJavaFormat() uses the function to
determine whether it should format a given document full or not, and
it makes it difficult to make the wrapper functions git-op free.

This CL simply sets opts.full if the default is set.
Note that this CL makes behaviour changes for Python.

1) depot_tools supports partial formatting only for Py, Java, and C++.
i.e., git cl format always formats the entire document unconditionally
for all the other languages, like rust, swift, and others.
Hence, this CL is noop for the other languages.

2) RunClangFormatDiff() and RunGoogleJavaFormat() calls
settings.GetFormatFullByDefault() to find the default format option
even if opts.full is False, as noted above.
Hence, this CL is noop for them.

3) RunYapf() doesn't honor settings.GetFormatFullByDefault().
It just looks at opts.full only to determine how to format the file.

Therefore, after this CL is introduced, git cl format will start formatting
python files with --full if 'rietveld.format-full-by-default' in the git
config is set.

--------
I think this is a resaonable change for consistency. That said,
It seems that 'rietveld.format-full-by-default' is set manually, and
it's possible that virtually nobody has the config.

If one wants to run git cl format with --full always, then they should
just use git alias, instead of the config.


Bug: 386840832
Change-Id: Ibd7210e20897d354bd3eddf84927cf22105b7255