git_cl: sets opts.full if format-full-by-default is set
settings.GetFormatFullByDefault() looks up a git config to find if
format-full-by-default is set.

RunClangFormatDiff() and RunGoogleJavaFormat() uses the function to
determine whether it should format a given document full or not, and
it makes it difficult to make the wrapper functions git-op free.

This CL simply sets opts.full if the default is set.

Bug: 386840832
Change-Id: Ibd7210e20897d354bd3eddf84927cf22105b7255