git_cl: move format launch functions into a separate module
presubmit_canned_checks.py:CheckPatchFormatted() runs
`git cl format --dry-run` to check the format of the changes.

When CiderG runs presubmit, it becomes an issue because git cl format
command runs a number of git commands which are not available in Cog.

The plan is to move the formatter _RunXX functions to a separate module
so that it will become easier to keep the implementations git-free and
presubmit_canned_checks.py will call a function in the module to
validate formats instead of running `git cl format --dry-run`.

This CL simply moves the _RunXXX functions to cl_format.py without
any meaningful changes. Follow-up CLs will refactor the codes to
execute the plan.

Change-Id: I5d610a73a62515c666667e97d70bb834e96cb072