metrics_xml_format: check the XML path against realpath
metrics_xml_format checks the given file path to ensure that
it's an XML under tools/metrics with the following steps.
1. convert the input file path as a real path.
2. check the real path starts with
  GetPrimarySolutionPath() + "/tools/metrics"

This doesn't work as intended if GetPrimarySolutionPath() returns
a symblink.

This patch converts GetPrimarySolutionPath() as a realpath to ensure
it runs the check based on the realpath of the git checkout.

Bug: 396182402
Change-Id: I3207eba7ed52d09a01b38d65fcf7c314622d95cb