[unify-auth] deprecate custom boto options download_for_google_storage
The end goal is that everything will be passed through to the wrapped
gsutil command where luci-auth can use an existing BOTO if configured
in the standard way but not for any auth.

One tangible outcome of this is that by removing the --no_auth option
we may be generating tokens and wrapping the command in the luci auth
context for no reason.

Overall it will simplify the use of download_for_google_storage and
prevent any non luci controlled BOTO_CONFIGS auth being used.

Change-Id: I8fcd20f2b03797da742fbf7b1db173a31f213fed