[3PWebAPIRequestFwd] Obsolete the ThirdPartyCredentialManagerFactory
ThirdPartyCredentialManagerFactory was supposed to be used to create an
instance of CredentialManagerInterface when 3rd party mode is used in
Chrome. The approach has changed in the meantime and the factory is no
longer needed. ContentCredentialManager will just receive a
CredentialManagerInterface when it is constructed, instead of using a
one-time-use factory.
This is a pre-work for: https://crrev.com/c/6372601.

Bug: 406462724
Change-Id: I1a6c266142003cf004fcc36ff2c1cfb0beca85af