set verobse=False for mirror.populate in _UpdateMirrorIfNotContains()
In https://crrev.com/c/6442501, apply_patch_ref() was updated to
refresh the git cache when applying patches in gclient. However,
it causes an excessive number of logs to be created, particularly
for src/v8, as the repo has a huge number of branches.

This CL simply sets the verbose option with False,
ignoring the value in options.verbose.

Bug: 407795715
Change-Id: Ibf32ff67d23f41b398cca82372c17d7ca331db26