git-cache: contains_revision returns False if sentient file exists
If git_cache.populate() gets interrupted midway, it can possibly leave
a git mirror in a state where some commits are available. Cloning such
a repo results in an empty repository and causes subsequent git
checkout to fail.

If the sentient file exists, it indicates that the bootstrapping
process was interrupted. This CL updates git_cache.contains_revision()
such that it returns False if the sentient file exists. Then,
the caller can call mirror.populate() to re-populate the repo/rev.

Bug: 41488725
Change-Id: I24fd24e300f5a9f9349589496d6b5f3dacf71fd2