lockfile: release the lock before closing the file handler
After crrev.com/c/6486489 landed, gclient sync failed randomly due to
lock failures. After investigation, it was concluded that the locks
are not immediately released after the lock file handler close so that
it fails to acquire the lock if gclient_scm.py attempts to lock it
immediately after mirror.populate().

In Linux, there is no guarantee that the file close will release
all the locks before processing the file description closure, unlike
the CloseHandle() do in Microsoft Windows.

This CL is to update the logic so that it releases the lock
before os.close().

Bug: 407795715
Change-Id: I0f58627d368922f27c0590dcea2e7fde4242ae17