git_cache: give 20 secs timeout to lockfile.Lock() in contains_revision
https://crrev.com/c/6507483 changed contains_revision() to acquire
the cache lock before checking the existence of the sentinel file
to avoid race conditions.

After the CL, git_cache.py acquires the lock in the following places.
1. sync()
2. contains_revision()

While it's reasonable to raise LockError() while there is an ongoing
sync(), the timeout can cause unexpected LockError() if multiple
processes are calling contains_revision() at the same time.

This CL simply gives it 20 secs, meaning that it would retry acquiring
the lock every 0.1s until it reaches 20 secs. A better approach is
using a read-write lock.

Bug: 421234614
Change-Id: I295652a2ef20a5b1f732908a522548c43c05d934