autoninja: Enable Siso by default for external developers
This CL changes the default build system for Chromium builds by external
developers from Ninja to Siso.
If Siso doesn't work for you, please file a bug and opt-out it.

How to file a bug:
(TODO: Prepare a Buganizer template for Siso user issues.)

How to opt-out Siso:
- Set `use_siso=false` in your `gn.args`.
- If you have `use_remoteexec=true`, `use_reclient=true` is also necessary.

How to revert this CL:
- Please revert this CL and https://crrev.com/c/6638830 on chromium/src together.

Bug: 409223168
Change-Id: I382746cab2e8b40db5f555711e965f814d089987