[gclient] Fix an issue with gclient hanging.
gclient hangs when I try to update my infra checkout,
running pstree reveals that we're waiting for less to
complete, which it can't do because of how gclient
invokes less.

In the absence of a more robust mechanism for making
git non-interactive, I vote we just set the pager to
cat and the editor to false and call it a day.

Change-Id: Icc66c0a91d45426bb5f65f2ffb325c4f6b6e532d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/7008969
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Gavin Mak <gavinmak@google.com>